spring_v1

+++ src/main/java/blinov_first/dao/MediaFileDao.java
@@ -2,12 +2,20 @@ package blinov_first.dao;
 
 import blinov_first.entity.MediaFile;
 import blinov_first.exception.DaoException;
+
 import java.util.List;
 import java.util.Optional;
 
+/**
+ * Data-access contract for {@link MediaFile} entities.
+ */
 public interface MediaFileDao {
-    Optional<MediaFile> findById(int id) throws DaoException;
+
+    Optional<MediaFile> findById(long id) throws DaoException;
+
     List<MediaFile> findByUserId(Long userId) throws DaoException;
+
     boolean add(MediaFile file) throws DaoException;
-    boolean deleteById(int id, Long userId) throws DaoException;
+
+    boolean deleteById(long id, Long userId) throws DaoException;
 }
